requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
psycopg2-binary
pydantic
python-dotenv
pydantic-settings
